moves getDataOnTxDigests to same directory as other tx table code
