+++ explorer/client/src/components/transaction-card/TxCardUtils.tsx
@@ -2,12 +2,25 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import {
+    getExecutionStatusType,
+    getTotalGasUsed,
+    getTransactions,
+    getTransactionDigest,
+    getTransactionKindName,
+    getTransferObjectTransaction,
+    getTransferSuiTransaction,
+    getTransferSuiAmount,
+    type GetTxnDigestsResponse,
+    type CertifiedTransaction,
     type ExecutionStatusType,
     type TransactionKindName,
 } from '@mysten/sui.js';
 import BN from 'bn.js';
 
+import { DefaultRpcClient } from '../../utils/api/DefaultRpcClient';
+import { type Network } from '../../utils/api/rpcSetting';
 import { numberSuffix } from '../../utils/numberUtil';
+import { deduplicate } from '../../utils/searchUtil';
 import { truncate, presentBN } from '../../utils/stringUtils';
 import { timeAgo } from '../../utils/timeUtils';
 
@@ -122,3 +135,56 @@ export const genTableDataFromTxData = (
         ],
     };
 };
+
+export const getDataOnTxDigests = (
+    network: Network | string,
+    transactions: GetTxnDigestsResponse
+) =>
+    DefaultRpcClient(network)
+        .getTransactionWithEffectsBatch(deduplicate(transactions))
+        .then((txEffs) => {
+            return (
+                txEffs
+                    .map((txEff) => {
+                        const [seq, digest] = transactions.filter(
+                            (transactionId) =>
+                                transactionId[1] ===
+                                getTransactionDigest(txEff.certificate)
+                        )[0];
+                        const res: CertifiedTransaction = txEff.certificate;
+                        // TODO: handle multiple transactions
+                        const txns = getTransactions(res);
+                        if (txns.length > 1) {
+                            console.error(
+                                'Handling multiple transactions is not yet supported',
+                                txEff
+                            );
+                            return null;
+                        }
+                        const txn = txns[0];
+                        const txKind = getTransactionKindName(txn);
+                        const recipient =
+                            getTransferObjectTransaction(txn)?.recipient ||
+                            getTransferSuiTransaction(txn)?.recipient;
+
+                        return {
+                            seq,
+                            txId: digest,
+                            status: getExecutionStatusType(txEff),
+                            txGas: getTotalGasUsed(txEff),
+                            suiAmount: getTransferSuiAmount(txn),
+                            kind: txKind,
+                            From: res.data.sender,
+                            timestamp_ms: txEff.timestamp_ms,
+                            ...(recipient
+                                ? {
+                                      To: recipient,
+                                  }
+                                : {}),
+                        };
+                    })
+                    // Remove failed transactions and sort by sequence number
+                    .filter((itm) => itm)
+                    .sort((a, b) => b!.seq - a!.seq)
+            );
+        });
